ST-1071 => Merging RequestResponseAspect.java into RestProviderHttpResponseCodeAspect.java
