--- src/test/java/gov/va/ascent/framework/audit/BaseAuditAspectTest.java
@@ -1,188 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package gov.va.ascent.framework.audit;
-
-import org.aspectj.lang.JoinPoint;
-import org.aspectj.lang.ProceedingJoinPoint;
-import org.aspectj.lang.reflect.MethodSignature;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.mockito.Mock;
-import org.mockito.runners.MockitoJUnitRunner;
-
-import java.lang.reflect.Method;
-
-import static org.junit.Assert.assertEquals;
-import static org.mockito.Mockito.when;
-
-/**
- *
- * @author rthota
- */
-@RunWith(MockitoJUnitRunner.class)
-public class BaseAuditAspectTest {
-
-    @Mock
-    private ProceedingJoinPoint proceedingJoinPoint;
-    @Mock
-    private MethodSignature signature;
-
-    
-    @Mock
-    private JoinPoint.StaticPart staticPart;
-    private Object[] value;
-
-    
-	@Before
-	public void setUp() throws Exception {
-		value = new Object[1];
-		value[0] = "";
-		when(proceedingJoinPoint.getArgs()).thenReturn(value);
-		when(proceedingJoinPoint.getStaticPart()).thenReturn(staticPart);
-        when(proceedingJoinPoint.getSignature()).thenReturn(signature);
-        when(signature.getMethod()).thenReturn(myMethod());
-	}
-    /**
-     * Test of auditableAnnotation method, of class BaseAuditAspect.
-     */
-    @Test
-    public void testAuditableAnnotation() {
-        System.out.println("auditableAnnotation");
-        BaseAuditAspect.auditableAnnotation();
-    }
-
-    /**
-     * Test of auditableExecution method, of class BaseAuditAspect.
-     */
-    @Test
-    public void testAuditableExecution() {
-        System.out.println("auditableExecution");
-        BaseAuditAspect.auditableExecution();
-    }
-
-    /**
-     * Test of auditRestController method, of class BaseAuditAspect.
-     */
-    @Test
-    public void testAuditRestController() {
-        System.out.println("auditRestController");
-        BaseAuditAspect.auditRestController();
-    }
-
-    /**
-     * Test of auditPublicServiceResponseRestMethod method, of class BaseAuditAspect.
-     */
-    @Test
-    public void testAuditPublicServiceResponseRestMethod() {
-        System.out.println("auditPublicServiceResponseRestMethod");
-        BaseAuditAspect.auditPublicServiceResponseRestMethod();
-    }
-
-    /**
-     * Test of getMethodAndArgumentsAsString method, of class BaseAuditAspect.
-     */
-    @Test
-    public void testGetMethodAndArgumentsAsString() {
-        System.out.println("getMethodAndArgumentsAsString");
-        BaseAuditAspect instance = new BaseAuditAspect();
-        String expResult = "someMethod()";
-        String result = instance.getMethodAndArgumentsAsString(proceedingJoinPoint);
-        assertEquals(expResult, result);
-
-    }
-
-    /**
-     * Test of getMethodName method, of class BaseAuditAspect.
-     */
-    @Test
-    public void testGetMethodName() {
-        System.out.println("getMethodName");
-        BaseAuditAspect instance = new BaseAuditAspect();
-        String expResult = "someMethod";
-        String result = instance.getMethodName(proceedingJoinPoint);
-        assertEquals(expResult, result);
-    }
-
-    /**
-     * Test of getResultAsString method, of class BaseAuditAspect.
-     */
-    @Test
-    public void testGetResultAsString() {
-        System.out.println("getResultAsString");
-        Object result_2 = "test";
-        BaseAuditAspect instance = new BaseAuditAspect();
-        String expResult = " returned test";
-        String result = instance.getResultAsString(result_2);
-        assertEquals(expResult, result);
-       
-    }
-
-    /**
-     * Test of getExceptionAsString method, of class BaseAuditAspect.
-     */
-    @Test
-    public void testGetExceptionAsString() {
-        System.out.println("getExceptionAsString");
-        Throwable ex = new Throwable();
-        long duration = 0L;
-        BaseAuditAspect instance = new BaseAuditAspect();
-        String expResult = " threw Throwable after 0 msecs with message null";
-        String result = instance.getExceptionAsString(ex, duration);
-        assertEquals(expResult, result);
-    }
-
-    /**
-     * Test of getDefaultAuditableInstance method, of class BaseAuditAspect.
-     */
-    @Test
-    public void testGetDefaultAuditableInstance() throws Exception {
-        System.out.println("getDefaultAuditableInstance");
-        Method method = myMethod();
-        AuditEvents expResult = AuditEvents.REQUEST_RESPONSE;
-        AuditEventData result = BaseAuditAspect.getDefaultAuditableInstance(method);
-        assertEquals(expResult, result.getEvent());
-        assertEquals("someMethod", result.getActivity());
-        assertEquals("gov.va.ascent.framework.audit.BaseAuditAspectTest", result.getAuditClass());
-    }
-
-    /**
-     * Test of getDefaultAuditableInstance method, of class BaseAuditAspect.
-     */
-    @Test
-    public void testGetDefaultAuditableInstanceMethodNotNull() throws Exception{
-        System.out.println("getDefaultAuditableInstance");
-        Method method = myMethod();
-        AuditEvents expResult = AuditEvents.REQUEST_RESPONSE;
-        AuditEventData result = BaseAuditAspect.getDefaultAuditableInstance(method);
-        assertEquals(expResult, result.getEvent());
-        assertEquals(method.getName(), result.getActivity());
-        assertEquals(method.getDeclaringClass().getName(), result.getAuditClass());
-    }
-
-    /**
-     * Test of getDefaultAuditableInstance method, of class BaseAuditAspect.
-     */
-    @Test
-    public void testGetDefaultAuditableInstanceMethodNull() throws Exception{
-        System.out.println("getDefaultAuditableInstance");
-        Method method = null;
-        AuditEvents expResult = AuditEvents.REQUEST_RESPONSE;
-        AuditEventData result = BaseAuditAspect.getDefaultAuditableInstance(method);
-        assertEquals(expResult, result.getEvent());
-        assertEquals("", result.getActivity());
-        assertEquals("", result.getAuditClass());
-    }
-    
-    public Method myMethod() throws NoSuchMethodException{
-        return getClass().getDeclaredMethod("someMethod");
-    }
-
-    public void someMethod() {
-        // do nothing
-    }
-    
-}
