+++ src/test/java/gov/va/ascent/framework/rest/provider/RestProviderHttpResponseCodeAspectTest.java
@@ -3,6 +3,7 @@ package gov.va.ascent.framework.rest.provider;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
+import static org.junit.Assert.assertTrue;
 import static org.mockito.Mockito.when;
 
 import java.lang.reflect.Method;
@@ -18,13 +19,19 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.mock.web.MockHttpServletRequest;
+import org.springframework.web.context.request.RequestContextHolder;
+import org.springframework.web.context.request.ServletRequestAttributes;
+import org.springframework.web.context.support.AnnotationConfigWebApplicationContext;
 
 import gov.va.ascent.framework.messages.Message;
 import gov.va.ascent.framework.messages.MessageSeverity;
+import gov.va.ascent.framework.service.ServiceRequest;
 import gov.va.ascent.framework.service.ServiceResponse;
 
 @RunWith(MockitoJUnitRunner.class)
@@ -45,13 +52,36 @@ public class RestProviderHttpResponseCodeAspectTest {
     @Mock
     private MethodSignature mockSignature;
     
+    
+	@InjectMocks
+	private RestProviderHttpResponseCodeAspect requestResponseAspect = new RestProviderHttpResponseCodeAspect();
+
+	private AnnotationConfigWebApplicationContext context;
+
+    private TestServiceRequest mockRequestObject = new TestServiceRequest();
+    private Object[] mockArray = {mockRequestObject};
+
+    
     private List<Message> detailedMsg = new ArrayList<Message>();
 
 	@Before
 	public void setUp() throws Exception {
+		MockHttpServletRequest httpServletRequest = new MockHttpServletRequest();
+		RequestContextHolder.setRequestAttributes(new ServletRequestAttributes(httpServletRequest));
+		
+		try{
+
+
+		}catch(Throwable e) {
+
+		}
 		RestProviderHttpResponseCodeAspectLogAppender.events.clear();
 		restProviderLog.setLevel(Level.DEBUG);
 		try {
+			when(proceedingJoinPoint.getArgs()).thenReturn(mockArray);
+			when(proceedingJoinPoint.getSignature()).thenReturn(mockSignature);
+			when(mockSignature.getMethod()).thenReturn(myMethod());
+			
 			Message msg = new Message(MessageSeverity.FATAL,"FatalKey","Fatal Message");
 			detailedMsg.add(msg);
 			when(proceedingJoinPoint.proceed()).thenReturn(responseEntity);
@@ -134,7 +164,45 @@ public class RestProviderHttpResponseCodeAspectTest {
 		assertEquals("gov.va.ascent.framework.exception.AscentRuntimeException", RestProviderHttpResponseCodeAspectLogAppender.events.get(0).getThrowableProxy().getClassName());
 	}
 	
-	
+	@Test
+	public void testAnnotatedMethodRequestResponse() {
+		Object obj;
+		try {
+			obj = restProviderHttpResponseCodeAspect.logAnnotatedMethodRequestResponse(proceedingJoinPoint);
+			assertNotNull(obj);
+		} catch (Throwable throwable) {
+			assertTrue(throwable instanceof RuntimeException);
+		}
+	}
+
+	@Test
+	public void testAnnotatedMethodRequestResponseRunTimeException() {
+
+		try {
+			Object[] array = {null, new Object()};
+			when(proceedingJoinPoint.getArgs()).thenReturn(array);
+			when(proceedingJoinPoint.proceed()).thenThrow(new RuntimeException("Unit Test Exception"));
+			restProviderHttpResponseCodeAspect.logAnnotatedMethodRequestResponse(proceedingJoinPoint);
+		} catch(Throwable throwable){
+			assertTrue(throwable instanceof RuntimeException);
+		}
+
+	}
+
+	@Test
+	public void testAnnotatedMethodRequestResponseRunTimeExceptionArrayZero() {
+
+		try {
+			Object[] array = new Object[0];
+			when(proceedingJoinPoint.getArgs()).thenReturn(array);
+			when(proceedingJoinPoint.proceed()).thenThrow(new RuntimeException("Unit Test Exception"));
+			restProviderHttpResponseCodeAspect.logAnnotatedMethodRequestResponse(proceedingJoinPoint);
+		} catch(Throwable throwable){
+			assertTrue(throwable instanceof RuntimeException);
+		}
+
+	}
+
     public Method myMethod() throws NoSuchMethodException{
         return getClass().getDeclaredMethod("someMethod");
     }
@@ -146,4 +214,16 @@ public class RestProviderHttpResponseCodeAspectTest {
 	class TestClass {
 		
 	}
+}
+
+class TestServiceRequest extends ServiceRequest {
+    private String text;
+
+    public String getText() {
+        return text;
+    }
+
+    public void setText(String text) {
+        this.text = text;
+    }	
 }
